Move client/userd/k8s to client/k8s
This commit performs no change of the logic. It's just moving stuff.

- client/userd/k8s is moved to client/k8s.
- k8s_cluster.go renamed to cluster.go
- k8s_outbound.go merged into cluster.go
